ペルソナコアまわりの責務分離リファクタを開始
