+++ persona/constants.py
@@ -1,0 +1,14 @@
+"""Shared configuration values for persona components."""
+from persona.utils import env_int
+
+
+RECALL_SNIPPET_MAX_CHARS = env_int("SAIVERSE_RECALL_SNIPPET_MAX_CHARS", 8000)
+RECALL_SNIPPET_STREAM_MAX_CHARS = env_int("SAIVERSE_RECALL_SNIPPET_STREAM_MAX_CHARS", 800)
+RECALL_SNIPPET_PULSE_MAX_CHARS = env_int("SAIVERSE_RECALL_SNIPPET_PULSE_MAX_CHARS", 1200)
+
+
+__all__ = [
+    "RECALL_SNIPPET_MAX_CHARS",
+    "RECALL_SNIPPET_STREAM_MAX_CHARS",
+    "RECALL_SNIPPET_PULSE_MAX_CHARS",
+]
