+++ persona/core.py
@@ -3,9 +3,8 @@ import logging
 import time
 import copy
 import os
-import html
 from pathlib import Path
-from typing import Any, Callable, Dict, List, Optional, Tuple, Iterator
+from typing import Any, Callable, Dict, List, Optional, Tuple
 from datetime import datetime, timezone as dt_timezone, tzinfo, timedelta
 from zoneinfo import ZoneInfo
 
@@ -16,7 +15,7 @@ from dotenv import load_dotenv
 
 from buildings import Building
 from saiverse_memory import SAIMemoryAdapter
-from llm_clients import get_llm_client, LLMClient
+from llm_clients import get_llm_client
 import llm_clients
 from model_configs import model_supports_images
 from action_handler import ActionHandler
@@ -26,19 +25,31 @@ from database.models import AI as AIModel
 from tools import TOOL_REGISTRY, TOOL_SCHEMAS
 from tools.defs import parse_tool_result
 from tools.context import persona_context
-from persona.bootstrap import initialise_memory_adapter, load_action_priority, load_session_data
+from persona.bootstrap import (
+    initialise_memory_adapter,
+    load_action_priority,
+    load_session_data,
+)
+from persona.constants import (
+    RECALL_SNIPPET_PULSE_MAX_CHARS,
+)
 from persona.history import initialise_pulse_state
-from persona.utils import env_int
+from persona.mixins import (
+    PersonaGenerationMixin,
+    PersonaHistoryMixin,
+    PersonaMovementMixin,
+    PersonaEmotionMixin,
+)
 
 load_dotenv()
 
 
-RECALL_SNIPPET_MAX_CHARS = env_int("SAIVERSE_RECALL_SNIPPET_MAX_CHARS", 8000)
-RECALL_SNIPPET_STREAM_MAX_CHARS = env_int("SAIVERSE_RECALL_SNIPPET_STREAM_MAX_CHARS", 800)
-RECALL_SNIPPET_PULSE_MAX_CHARS = env_int("SAIVERSE_RECALL_SNIPPET_PULSE_MAX_CHARS", 1200)
-
-
-class PersonaCore:
+class PersonaCore(
+    PersonaGenerationMixin,
+    PersonaHistoryMixin,
+    PersonaMovementMixin,
+    PersonaEmotionMixin,
+):
     def __init__(
         self,
         city_name: str,
@@ -160,924 +171,6 @@ class PersonaCore:
         self.pending_attachment_metadata: List[Dict[str, Any]] = []
 
 
-
-    def _parse_timestamp_to_utc(self, value: Any) -> Optional[datetime]:
-        """Convert various timestamp formats to aware UTC datetime."""
-        if value is None:
-            return None
-        try:
-            if isinstance(value, (int, float)):
-                return datetime.fromtimestamp(float(value), tz=dt_timezone.utc)
-            if isinstance(value, str):
-                raw = value.strip()
-                if not raw:
-                    return None
-                if raw.endswith("Z"):
-                    raw = raw[:-1] + "+00:00"
-                parsed = datetime.fromisoformat(raw)
-                if parsed.tzinfo is None:
-                    return parsed.replace(tzinfo=dt_timezone.utc)
-                return parsed.astimezone(dt_timezone.utc)
-        except Exception:
-            logging.debug("Failed to parse timestamp '%s' for persona %s", value, self.persona_id)
-            return None
-        return None
-
-    def _format_timezone_offset(self, dt_obj: datetime) -> str:
-        """Return a formatted UTC offset string."""
-        offset = dt_obj.utcoffset()
-        if offset is None:
-            return "UTC+00:00"
-        total_minutes = int(offset.total_seconds() // 60)
-        sign = "+" if total_minutes >= 0 else "-"
-        total_minutes = abs(total_minutes)
-        hours, minutes = divmod(total_minutes, 60)
-        return f"UTC{sign}{hours:02d}:{minutes:02d}"
-
-    def _format_local_timestamp(self, dt_obj: datetime) -> str:
-        """Format a localized timestamp string for recent history."""
-        offset_label = self._format_timezone_offset(dt_obj)
-        tz_label = dt_obj.tzname() or self.timezone_name
-        return f"{dt_obj.strftime('%Y-%m-%d %H:%M:%S')} {tz_label} ({offset_label})"
-
-    def _format_elapsed(self, delta: timedelta) -> str:
-        """Convert a timedelta into a human-friendly Japanese string."""
-        if delta.total_seconds() < 0:
-            delta = timedelta(0)
-        total_minutes = int(delta.total_seconds() // 60)
-        days, rem_minutes = divmod(total_minutes, 1440)
-        hours, minutes = divmod(rem_minutes, 60)
-        parts: List[str] = []
-        if days:
-            parts.append(f"{days}日")
-        if hours:
-            parts.append(f"{hours}時間")
-        if minutes:
-            parts.append(f"{minutes}分")
-        if not parts:
-            parts.append("0分")
-        return " ".join(parts)
-
-    def _timestamp_to_epoch(self, primary: Any, secondary: Any = None) -> Optional[int]:
-        """
-        Convert timestamp-like values into epoch seconds.
-        `primary` is preferred (typically created_at), while `secondary` is used as fallback (typically timestamp).
-        """
-        result: Optional[int] = None
-        for candidate in (primary, secondary):
-            if candidate is None:
-                continue
-            if isinstance(candidate, (int, float)):
-                result = int(candidate)
-                break
-            if isinstance(candidate, str):
-                raw = candidate.strip()
-                if not raw:
-                    continue
-                try:
-                    result = int(float(raw))
-                    break
-                except ValueError:
-                    dt_obj = self._parse_timestamp_to_utc(raw)
-                    if dt_obj is not None:
-                        result = int(dt_obj.timestamp())
-                        break
-        logging.debug(
-            "[recall] normalised timestamp primary=%s secondary=%s -> %s",
-            primary,
-            secondary,
-            result,
-        )
-        return result
-
-
-    def _occupants_snapshot(self, building_id: str) -> List[str]:
-        occupants = self.occupants.get(building_id, []) or []
-        snapshot = []
-        for pid in occupants:
-            if not pid:
-                continue
-            pid_str = str(pid)
-            if pid_str not in snapshot:
-                snapshot.append(pid_str)
-        if building_id == self.current_building_id and self.persona_id not in snapshot:
-            snapshot.append(self.persona_id)
-        return snapshot
-
-    def _save_session_metadata(self) -> None:
-        """ペルソナの動的な状態をDBに保存し、各種ログファイルを保存する"""
-        if self.is_visitor:
-            # Visitors only save file-based logs, not DB state.
-            self.history_manager.save_all()
-            self._save_conscious_log()
-            return
-
-        db = self.SessionLocal()
-        try:
-            update_data = {
-                "EMOTION": json.dumps(self.emotion, ensure_ascii=False),
-                "AUTO_COUNT": self.auto_count,
-                "LAST_AUTO_PROMPT_TIMES": json.dumps(self.last_auto_prompt_times, ensure_ascii=False)
-            }
-            db.query(AIModel).filter(AIModel.AIID == self.persona_id).update(update_data)
-            db.commit()
-            logging.info(f"Saved dynamic state to DB for {self.persona_name}.")
-        except Exception as e:
-            db.rollback()
-            logging.error(f"Failed to save session data to DB for {self.persona_name}: {e}", exc_info=True)
-        finally:
-            db.close()
-        # self.messagesが空の場合はlog.jsonを上書きしない
-        if self.messages:
-            self.history_manager.save_all()
-        self._save_conscious_log()
-
-    def set_model(self, model: str, context_length: int, provider: str) -> None:
-        self.model = model
-        self.context_length = context_length
-        self.model_supports_images = model_supports_images(model)
-        self.llm_client = get_llm_client(model, provider, context_length)
-
-    def _build_messages(
-        self,
-        user_message: Optional[str],
-        extra_system_prompt: Optional[str] = None,
-        info_text: Optional[str] = None,
-        guidance_text: Optional[str] = None,
-        user_metadata: Optional[Dict[str, Any]] = None,
-    ) -> List[Dict[str, Any]]:
-        building = self.buildings[self.current_building_id]
-        now_local = datetime.now(self.timezone)
-        current_time = now_local.strftime("%H:%M")
-        system_text = self.common_prompt.format(
-            current_building_name=building.name,
-            current_building_system_instruction=building.system_instruction.format(current_time=current_time),
-            current_persona_id=self.persona_id,
-            current_persona_name=self.persona_name,
-            current_persona_system_instruction=self.persona_system_instruction,
-            current_time=current_time,
-            current_city_name=self.city_name,
-        )
-        emotion_text = self.emotion_prompt.format(
-            stability_mean=self.emotion["stability"]["mean"],
-            stability_var=self.emotion["stability"]["variance"],
-            affect_mean=self.emotion["affect"]["mean"],
-            affect_var=self.emotion["affect"]["variance"],
-            resonance_mean=self.emotion["resonance"]["mean"],
-            resonance_var=self.emotion["resonance"]["variance"],
-            attitude_mean=self.emotion["attitude"]["mean"],
-            attitude_var=self.emotion["attitude"]["variance"],
-        )
-        system_text = system_text + "\n" + emotion_text
-        if info_text:
-            system_text += (
-                "\n\n## 追加情報\n"
-                "常時稼働モジュールから以下の情報が提供されています。今回の発話にこの情報を利用してください。\n"
-                f"{info_text}"
-            )
-
-        base_chars = len(system_text)
-        if extra_system_prompt:
-            base_chars += len(extra_system_prompt)
-        if user_message:
-            base_chars += len(user_message)
-
-        history_limit = max(0, self.context_length - base_chars)
-        history_msgs = self.history_manager.get_recent_history(history_limit)
-        logging.debug(
-            "history_limit=%s context=%s base=%s history_count=%s",
-            history_limit,
-            self.context_length,
-            base_chars,
-            len(history_msgs),
-        )
-        if history_msgs:
-            logging.debug("history_head=%s", history_msgs[0])
-            logging.debug("history_tail=%s", history_msgs[-1])
-
-        sanitized_history: List[Dict[str, Any]] = []
-        for m in history_msgs:
-            role = m.get("role", "")
-            content = m.get("content", "")
-            if role == "system" and "### 意識モジュールからの情報提供" in content:
-                continue
-            sanitized: Dict[str, Any] = {"role": role, "content": content}
-            metadata = m.get("metadata")
-            if isinstance(metadata, dict):
-                sanitized["metadata"] = copy.deepcopy(metadata)
-            sanitized_history.append(sanitized)
-
-        msgs = [{"role": "system", "content": system_text}] + sanitized_history
-        if guidance_text:
-            wrapped_guidance = f"<system>\n{guidance_text}\n</system>"
-            msgs.append({"role": "user", "content": wrapped_guidance})
-        if extra_system_prompt:
-            msgs.append({"role": "system", "content": extra_system_prompt})
-        if user_message:
-            user_entry: Dict[str, Any] = {"role": "user", "content": user_message}
-            if isinstance(user_metadata, dict):
-                user_entry["metadata"] = copy.deepcopy(user_metadata)
-                logging.debug("[persona_core] user message metadata keys=%s", list(user_metadata.keys()))
-            msgs.append(user_entry)
-        return msgs
-
-    def _collect_recent_memory_timestamps(self) -> List[int]:
-        recent = self.history_manager.get_recent_history(self.context_length)
-        values: List[int] = []
-        seen = set()
-        for msg in recent:
-            value = self._timestamp_to_epoch(msg.get("created_at"), msg.get("timestamp"))
-            if value is None:
-                continue
-            if value in seen:
-                continue
-            seen.add(value)
-            values.append(value)
-        logging.debug(
-            "[recall] collected recent timestamps count=%d values=%s",
-            len(values),
-            values,
-        )
-        return values
-
-    def _process_generation_result(
-        self,
-        content: str,
-        user_message: Optional[str],
-        system_prompt_extra: Optional[str],
-        log_extra_prompt: bool = True,
-        user_metadata: Optional[Dict[str, Any]] = None,
-    ) -> Tuple[str, Optional[Dict[str, str]], bool]:
-        say, actions = self.action_handler.parse_response(content)
-        move_target, _, delta = self.action_handler.execute_actions(actions)
-
-        # --- New part for exploration ---
-        explore_target = None
-        # Find the first 'explore_city' action in the list
-        for action in actions:
-            if action.get("action") == "explore_city":
-                explore_target = {"city_id": action.get("city_id")}
-                break
-
-        # --- New part for creation ---
-        creation_target = None
-        for action in actions:
-            if action.get("action") == "create_persona":
-                creation_target = {
-                    "name": action.get("name"), "system_prompt": action.get("system_prompt")
-                }
-                break # Only handle one creation at a time
-
-        if delta:
-            self._apply_emotion_delta(delta)
-
-        prompt_text = user_message if user_message is not None else system_prompt_extra or ""
-        module_delta = self.emotion_module.evaluate(
-            prompt_text, say, current_emotion=self.emotion
-        )
-        if module_delta:
-            self._apply_emotion_delta(module_delta)
-
-        if system_prompt_extra and log_extra_prompt:
-            self.history_manager.add_message(
-                {"role": "user", "content": system_prompt_extra},
-                self.current_building_id,
-                heard_by=self._occupants_snapshot(self.current_building_id),
-            )
-        if user_message:
-            user_entry: Dict[str, Any] = {"role": "user", "content": user_message}
-            if isinstance(user_metadata, dict):
-                user_entry["metadata"] = copy.deepcopy(user_metadata)
-            self.history_manager.add_message(
-                user_entry,
-                self.current_building_id,
-                heard_by=self._occupants_snapshot(self.current_building_id),
-            )
-
-        reasoning_entries = self.llm_client.consume_reasoning()
-        attachments_from_llm = self.llm_client.consume_attachments()
-        combined_media: List[Dict[str, Any]] = []
-        for meta in attachments_from_llm + self.pending_attachment_metadata:
-            if not isinstance(meta, dict):
-                continue
-            media_items = meta.get("media")
-            if not isinstance(media_items, list):
-                continue
-            for item in media_items:
-                if isinstance(item, dict):
-                    combined_media.append(copy.deepcopy(item))
-        self.pending_attachment_metadata = []
-        metadata_payload: Optional[Dict[str, Any]] = None
-        if combined_media:
-            metadata_payload = {"media": combined_media}
-
-        building_content = self._combine_with_reasoning(content, reasoning_entries)
-        persona_msg: Dict[str, Any] = {"role": "assistant", "content": content}
-        building_msg: Dict[str, Any] = {"role": "assistant", "content": building_content}
-        if metadata_payload:
-            persona_msg["metadata"] = metadata_payload
-            building_msg["metadata"] = metadata_payload
-        self.history_manager.add_to_persona_only(persona_msg)
-        self.history_manager.add_to_building_only(
-            self.current_building_id,
-            building_msg,
-            heard_by=self._occupants_snapshot(self.current_building_id),
-        )
-
-        moved = self._handle_movement(move_target)
-        self._handle_exploration(explore_target) # Call the new handler
-        self._handle_creation(creation_target) # Call the creation handler
-        self._save_session_metadata()
-        return say, move_target, moved
-
-    def _generate(
-        self,
-        user_message: Optional[str],
-        user_metadata: Optional[Dict[str, Any]] = None,
-        system_prompt_extra: Optional[str] = None,
-        info_text: Optional[str] = None,
-        guidance_text_override: Optional[str] = None,
-        log_extra_prompt: bool = True,
-        log_user_message: bool = True,
-    ) -> tuple[str, Optional[Dict[str, str]], bool]:
-        prev_emotion_state = copy.deepcopy(self.emotion)
-        actual_user_message = user_message
-        if user_message is None and system_prompt_extra is None:
-            history = self.history_manager.building_histories.get(
-                self.current_building_id, []
-            )
-            if not history or history[-1].get("role") != "user":
-                actual_user_message = "意識モジュールが発話することを意思決定しました。自由に発言してください"
-                logging.debug("Injected user message for context")
-
-        # Memory: prepare recall context from SAIMemory
-        combined_info = info_text or ""
-        if self.sai_memory is not None:
-            try:
-                recall_source = user_message.strip() if (user_message and user_message.strip()) else None
-                if recall_source is None:
-                    recall_source = self.history_manager.get_last_user_message()
-                if recall_source:
-                    exclude_times = self._collect_recent_memory_timestamps()
-                    logging.debug(
-                        "[recall] invoking recall_snippet building=%s source_preview=%s exclude_times=%s",
-                        self.current_building_id,
-                        recall_source[:120],
-                        exclude_times,
-                    )
-                    snippet = self.sai_memory.recall_snippet(
-                        self.current_building_id,
-                        recall_source,
-                        max_chars=RECALL_SNIPPET_MAX_CHARS,
-                        exclude_created_at=exclude_times,
-                    )
-                    if snippet:
-                        logging.debug("[memory] recall snippet content=%s", snippet[:400])
-                        combined_info = (combined_info + "\n" + snippet).strip() if combined_info else snippet
-                        logging.debug("[memory] SAIMemory recall snippet included (%d chars)", len(snippet))
-            except Exception as exc:
-                logging.warning("SAIMemory recall failed: %s", exc)
-
-        msgs = self._build_messages(
-            actual_user_message,
-            extra_system_prompt=system_prompt_extra,
-            info_text=combined_info or None,
-            guidance_text=guidance_text_override,
-            user_metadata=user_metadata if actual_user_message == user_message else None,
-        )
-        logging.debug("Messages sent to API: %s", msgs)
-
-        content = self.llm_client.generate(msgs, tools=[])
-        attempt = 1
-        while content.strip() == "エラーが発生しました。" and attempt < 3:
-            logging.warning(
-                "LLM generation failed; retrying in 10s (%d/3)", attempt
-            )
-            time.sleep(10)
-            content = self.llm_client.generate(msgs, tools=[])
-            attempt += 1
-
-        logging.info("AI Response :\n%s", content)
-        say, move_target, changed = self._process_generation_result(
-            content,
-            user_message if log_user_message else None,
-            system_prompt_extra,
-            log_extra_prompt,
-            user_metadata if log_user_message else None,
-        )
-        self._post_response_updates(
-            prev_emotion_state,
-            user_message,
-            system_prompt_extra,
-            say,
-        )
-        return say, move_target, changed
-
-    def _generate_stream(
-        self,
-        user_message: Optional[str],
-        user_metadata: Optional[Dict[str, Any]] = None,
-        system_prompt_extra: Optional[str] = None,
-        info_text: Optional[str] = None,
-        guidance_text_override: Optional[str] = None,
-        log_extra_prompt: bool = True,
-        log_user_message: bool = True,
-    ) -> Iterator[str]:
-        prev_emotion_state = copy.deepcopy(self.emotion)
-        actual_user_message = user_message
-        if user_message is None and system_prompt_extra is None:
-            history = self.history_manager.building_histories.get(
-                self.current_building_id, []
-            )
-            if not history or history[-1].get("role") != "user":
-                actual_user_message = "意識モジュールが発話することを意思決定しました。自由に発言してください"
-                logging.debug("Injected user message for context")
-
-        # Memory: prepare recall context from SAIMemory
-        combined_info = info_text or ""
-        if self.sai_memory is not None:
-            try:
-                recall_source = user_message.strip() if (user_message and user_message.strip()) else None
-                if recall_source is None:
-                    recall_source = self.history_manager.get_last_user_message()
-                if recall_source:
-                    exclude_times = self._collect_recent_memory_timestamps()
-                    logging.debug(
-                        "[recall] invoking recall_snippet(stream) building=%s source_preview=%s exclude_times=%s",
-                        self.current_building_id,
-                        recall_source[:120],
-                        exclude_times,
-                    )
-                    snippet = self.sai_memory.recall_snippet(
-                        self.current_building_id,
-                        recall_source,
-                        max_chars=RECALL_SNIPPET_STREAM_MAX_CHARS,
-                        exclude_created_at=exclude_times,
-                    )
-                    if snippet:
-                        combined_info = (combined_info + "\n" + snippet).strip() if combined_info else snippet
-            except Exception as exc:
-                logging.warning("SAIMemory recall failed: %s", exc)
-
-        msgs = self._build_messages(
-            actual_user_message,
-            extra_system_prompt=system_prompt_extra,
-            info_text=combined_info or None,
-            guidance_text=guidance_text_override,
-            user_metadata=user_metadata if actual_user_message == user_message else None,
-        )
-        logging.debug("Messages sent to API: %s", msgs)
-
-        attempt = 1
-        while True:
-            content_accumulator = ""
-            tokens: List[str] = []
-            for token in self.llm_client.generate_stream(msgs, tools=[]):
-                content_accumulator += token
-                tokens.append(token)
-
-            if content_accumulator.strip() != "エラーが発生しました。" or attempt >= 3:
-                for t in tokens:
-                    yield t
-                break
-
-            logging.warning(
-                "LLM stream generation failed; retrying in 10s (%d/3)", attempt
-            )
-            attempt += 1
-            time.sleep(10)
-
-        logging.info("AI Response :\n%s", content_accumulator)
-        say, move_target, changed = self._process_generation_result(
-            content_accumulator,
-            user_message if log_user_message else None,
-            system_prompt_extra,
-            log_extra_prompt,
-            user_metadata if log_user_message else None,
-        )
-        self._post_response_updates(
-            prev_emotion_state,
-            user_message,
-            system_prompt_extra,
-            say,
-        )
-        return (say, move_target, changed)
-
-    def _handle_movement(self, move_target: Optional[Dict[str, str]]) -> bool:
-        if not move_target or not move_target.get("building"):
-            return False
-
-        prev_id = self.current_building_id
-        target_building_id = move_target.get("building")
-        target_city_id = move_target.get("city")
-        moved = False
-
-        # City間移動の処理
-        if target_city_id:
-            if self.dispatch_callback:
-                logging.info(f"Attempting to dispatch to city: {target_city_id}, building: {target_building_id}")
-                dispatched, reason = self.dispatch_callback(self.persona_id, target_city_id, target_building_id)
-                if dispatched:
-                    # 派遣が成功した場合、このインスタンスはまもなく削除されるため、
-                    # これ以上の状態変更は行わない。
-                    # 'moved' はTrue とみなし、後続の自律会話ループなどを抑制する。
-                    moved = True
-                else:
-                    logging.warning(f"Dispatch failed: {reason}")
-                    self.history_manager.add_message(
-                        {"role": "system", "content": f"別のCityへの移動に失敗しました。{reason}"},
-                        self.current_building_id,
-                        heard_by=self._occupants_snapshot(self.current_building_id),
-                    )
-            else:
-                logging.error("Dispatch callback is not set. Cannot move between cities.")
-                self.history_manager.add_message(
-                    {"role": "system", "content": "別のCityへ移動する機能が設定されていません。"},
-                    self.current_building_id,
-                    heard_by=self._occupants_snapshot(self.current_building_id),
-                )
-            return moved
-
-        # City内移動の処理
-        if target_building_id and target_building_id in self.buildings:
-            allowed, reason = True, None
-            if self.move_callback:
-                allowed, reason = self.move_callback(self.persona_id, self.current_building_id, target_building_id)
-            if allowed:
-                logging.info("Moving to building: %s", target_building_id)
-                self.current_building_id = target_building_id
-                # Mark entry point for perception windowing
-                self._mark_entry(self.current_building_id)
-                moved = True
-            else:
-                logging.info("Move blocked to building: %s", target_building_id)
-                self.history_manager.add_message(
-                    {"role": "system", "content": f"移動できませんでした。{reason}"},
-                    self.current_building_id,
-                    heard_by=self._occupants_snapshot(self.current_building_id),
-                )
-        elif target_building_id:
-            logging.info("Unknown building id received: %s, staying at %s", target_building_id, self.current_building_id)
-
-        if moved:
-            """self.auto_count = 0
-            if prev_id != self.user_room_id and self.current_building_id == self.user_room_id:
-                self.history_manager.add_to_building_only(
-                    self.user_room_id,
-                    {
-                        "role": "assistant",
-                        "content": f'<div class="note-box">🏢 Building:<br><b>{self.persona_name}が入室しました</b></div>',
-                    },
-                )
-            elif prev_id == self.user_room_id and self.current_building_id != self.user_room_id:
-                dest_name = self.buildings[self.current_building_id].name
-                self.history_manager.add_to_building_only(
-                    self.user_room_id,
-                    {
-                        "role": "assistant",
-                        "content": f'<div class="note-box">🏢 Building:<br><b>{self.persona_name}が{dest_name}に向かいました</b></div>',
-                    },
-                )"""
-        return moved
-
-    def _handle_exploration(self, explore_target: Optional[Dict[str, str]]) -> None:
-        """Handles the 'explore_city' action by invoking the callback."""
-        if not explore_target or not explore_target.get("city_id"):
-            return
-
-        target_city_id = explore_target.get("city_id")
-        
-        if self.explore_callback:
-            logging.info(f"Attempting to explore city: {target_city_id}")
-            # The callback will handle API calls and feedback to the user.
-            self.explore_callback(self.persona_id, target_city_id)
-        else:
-            logging.error("Explore callback is not set. Cannot explore cities.")
-            self.history_manager.add_message(
-                {"role": "system", "content": "他のCityを探索する機能が設定されていません。"},
-                self.current_building_id,
-                heard_by=self._occupants_snapshot(self.current_building_id),
-            )
-
-    def _handle_creation(self, creation_target: Optional[Dict[str, str]]) -> None:
-        """Handles the 'create_persona' action by invoking the callback."""
-        if not creation_target or not creation_target.get("name") or not creation_target.get("system_prompt"):
-            return
-
-        name = creation_target.get("name")
-        system_prompt = creation_target.get("system_prompt")
-        
-        if self.create_persona_callback:
-            logging.info(f"Attempting to create persona: {name}")
-            success, message = self.create_persona_callback(name, system_prompt)
-            
-            # Provide feedback to the user/AI
-            feedback_message = f'<div class="note-box">🧬 ペルソナ創造:<br><b>{message}</b></div>'
-            self.history_manager.add_message(
-                {"role": "host", "content": feedback_message},
-                self.current_building_id,
-                heard_by=self._occupants_snapshot(self.current_building_id),
-            )
-        else:
-            logging.error("Create persona callback is not set. Cannot create new persona.")
-            self.history_manager.add_message(
-                {"role": "system", "content": "新しいペルソナを創造する機能が設定されていません。"},
-                self.current_building_id,
-                heard_by=self._occupants_snapshot(self.current_building_id),
-            )
-
-    def run_auto_conversation(self, initial: bool = False) -> List[str]:
-        replies: List[str] = []
-        move_target: Optional[Dict[str, str]] = None
-        building = self.buildings[self.current_building_id]
-        if initial and building.entry_prompt:
-            if building.run_entry_llm:
-                entry_text = building.entry_prompt.format(persona_name=self.persona_name)
-                say, move_target, _ = self._generate(None, system_prompt_extra=entry_text)
-                replies.append(say)
-            else:
-                self.history_manager.add_message(
-                    {"role": "system", "content": building.entry_prompt},
-                    self.current_building_id,
-                    heard_by=self._occupants_snapshot(self.current_building_id),
-                )
-        while (
-            building.auto_prompt
-            and building.run_auto_llm
-            and self.current_building_id == building.building_id
-            and (move_target is None or move_target.get("building") == building.building_id)
-            and self.auto_count < 10
-        ):
-            self.auto_count += 1
-            auto_text = building.auto_prompt.format(persona_name=self.persona_name)
-            say, move_target, changed = self._generate(None, system_prompt_extra=auto_text)
-            replies.append(say)
-            if changed:
-                building = self.buildings[self.current_building_id]
-                replies.extend(self.run_auto_conversation(initial=True))
-                break
-        return replies
-
-    def run_scheduled_prompt(self) -> List[str]:
-        building = self.buildings[self.current_building_id]
-        interval = getattr(building, "auto_interval_sec", 0)
-        if not (building.auto_prompt and building.run_auto_llm and interval > 0):
-            return []
-        last = self.last_auto_prompt_times.get(self.current_building_id, 0)
-        now = time.time()
-        if now - last < interval:
-            return []
-        self.last_auto_prompt_times[self.current_building_id] = now
-        auto_text = building.auto_prompt.format(persona_name=self.persona_name)
-        say, move_target, changed = self._generate(None, system_prompt_extra=auto_text)
-        replies = [say]
-        if changed:
-            replies.extend(self.run_auto_conversation(initial=True))
-        return replies
-
-    def handle_user_input(self, message: str, metadata: Optional[Dict[str, Any]] = None) -> List[str]:
-        logging.info("User input: %s", message)
-        if metadata:
-            logging.debug("[persona_core] handle_user_input received metadata keys=%s", list(metadata.keys()))
-        say, move_target, changed = self._generate(message, user_metadata=metadata)
-        replies = [say]
-
-        # This part remains to handle auto-conversation after a user-triggered one.
-        building = self.buildings[self.current_building_id]
-        if changed:
-            replies.extend(self.run_auto_conversation(initial=True))
-        elif (
-            building.auto_prompt
-            and building.run_auto_llm
-            and (move_target is None or move_target.get("building") == building.building_id)
-        ):
-            replies.extend(self.run_auto_conversation(initial=False))
-        return replies
-
-    def handle_user_input_stream(self, message: str, metadata: Optional[Dict[str, Any]] = None) -> Iterator[str]:
-        logging.info("User input: %s", message)
-        if metadata:
-            logging.debug("[persona_core] handle_user_input_stream received metadata keys=%s", list(metadata.keys()))
-
-        # ユーザーのメッセージを先に履歴に追加
-        user_entry: Dict[str, Any] = {"role": "user", "content": message}
-        if isinstance(metadata, dict):
-            user_entry["metadata"] = metadata
-        self.history_manager.add_message(
-            user_entry,
-            self.current_building_id,
-            heard_by=self._occupants_snapshot(self.current_building_id),
-        )
-
-        # _generate_stream には user_message=None を渡す
-        # これにより、_build_messages は履歴の最後のメッセージ（今追加したユーザーメッセージ）を文脈として使う
-        # また、_process_generation_result でユーザーメッセージが二重に記録されるのを防ぐ
-        gen = self._generate_stream(user_message=message, user_metadata=metadata, log_user_message=False)
-
-        try:
-            while True:
-                yield next(gen)
-        except StopIteration as e:
-            _, move_target, changed = e.value
-
-        building = self.buildings[self.current_building_id]
-        extra_replies: List[str] = []
-        if changed:
-            extra_replies.extend(self.run_auto_conversation(initial=True))
-        elif (
-            building.auto_prompt
-            and building.run_auto_llm
-            and (move_target is None or move_target.get("building") == building.building_id)
-        ):
-            extra_replies.extend(self.run_auto_conversation(initial=False))
-        for r in extra_replies:
-            yield r
-
-    def summon_to_user_room(self) -> List[str]:
-        prev = self.current_building_id
-        if prev == self.user_room_id:
-            return []
-        allowed, reason = True, None
-        if self.move_callback:
-            allowed, reason = self.move_callback(self.persona_id, self.current_building_id, self.user_room_id)
-        if not allowed:
-            self.history_manager.add_to_building_only(
-                self.user_room_id,
-                {"role": "assistant", "content": f'<div class="note-box">移動できませんでした。{reason}</div>'},
-                heard_by=self._occupants_snapshot(self.user_room_id),
-            )
-            self._save_session_metadata()
-            return []
-        self.current_building_id = self.user_room_id
-        self.auto_count = 0
-        # Mark entry into user room after switching
-        self._mark_entry(self.current_building_id)
-        self.history_manager.add_to_building_only(
-            self.user_room_id,
-            {
-                "role": "assistant",
-                "content": f'<div class="note-box">🏢 Building:<br><b>{self.persona_name}が入室しました</b></div>',
-            },
-            heard_by=self._occupants_snapshot(self.user_room_id),
-        )
-        self._save_session_metadata()
-        return self.run_auto_conversation(initial=True)
-
-    def get_building_history(self, building_id: str, raw: bool = False) -> List[Dict[str, str]]:
-        return self.history_manager.building_histories.get(building_id, [])
-
-    def _apply_emotion_delta(self, delta: Optional[List[Dict[str, Dict[str, float]]]]) -> None:
-        if not delta:
-            return
-        if isinstance(delta, dict):
-            delta = [delta]
-
-        for item in delta:
-            if not isinstance(item, dict):
-                continue
-            for key, val in item.items():
-                if key not in self.emotion:
-                    continue
-                if not isinstance(val, dict):
-                    continue
-
-                mean_delta = val.get("mean", 0)
-                var_delta = val.get("variance", 0)
-
-                try:
-                    mean_delta = float(mean_delta)
-                    var_delta = float(var_delta)
-                except (ValueError, TypeError):
-                    continue
-
-                current = self.emotion[key]
-                current["mean"] = max(-100.0, min(100.0, current["mean"] + mean_delta))
-                current["variance"] = max(0.0, min(100.0, current["variance"] + var_delta))
-
-
-    def _format_emotion_summary(self, prev: Dict[str, Dict[str, float]]) -> str:
-        labels = {
-            "stability": "安定性",
-            "affect": "情動",
-            "resonance": "共鳴",
-            "attitude": "態度",
-        }
-        lines = []
-        for key, label in labels.items():
-            before = prev.get(key, {"mean": 0.0, "variance": 1.0})
-            after = self.emotion.get(key, {"mean": 0.0, "variance": 1.0})
-            mean_delta = after["mean"] - before.get("mean", 0.0)
-            var_delta = after["variance"] - before.get("variance", 1.0)
-            line = (
-                f"{label}: mean {mean_delta:+.1f} → {after['mean']:.1f}, "
-                f"var {var_delta:+.1f} → {after['variance']:.1f}"
-            )
-            lines.append(line)
-        return '<div class="note-box">感情パラメータ変動<br>' + '<br>'.join(lines) + '</div>'
-
-    def _post_response_updates(
-        self,
-        prev_emotion: Dict[str, Dict[str, float]],
-        user_message: Optional[str],
-        system_prompt_extra: Optional[str],
-        assistant_message: str,
-    ) -> None:
-        prompt_text = ""
-        if user_message is not None:
-            prompt_text = user_message
-        elif system_prompt_extra:
-            prompt_text = system_prompt_extra
-
-        try:
-            module_delta = self.emotion_module.evaluate(
-                prompt_text,
-                assistant_message,
-                current_emotion=self.emotion,
-            )
-        except Exception:
-            logging.exception("[emotion] evaluation failed during post response update")
-            module_delta = None
-
-        if module_delta:
-            try:
-                self._apply_emotion_delta(module_delta)
-            except Exception:
-                logging.exception("[emotion] failed to apply module delta")
-
-        summary = self._format_emotion_summary(prev_emotion)
-        self.history_manager.add_to_persona_only({"role": "system", "content": summary})
-        self.history_manager.add_to_building_only(
-            self.current_building_id,
-            {"role": "assistant", "content": summary},
-            heard_by=self._occupants_snapshot(self.current_building_id),
-        )
-
-    def _combine_with_reasoning(self, base_text: str, reasoning_entries: List[Dict[str, str]]) -> str:
-        if not reasoning_entries:
-            return base_text
-
-        blocks: List[str] = []
-        for idx, entry in enumerate(reasoning_entries, start=1):
-            text = (entry.get("text") or "").strip()
-            if not text:
-                continue
-            title = (entry.get("title") or "").strip() or f"Thought {idx}"
-            safe_title = html.escape(title)
-            safe_text = html.escape(text).replace("\n", "<br>")
-            blocks.append(
-                f"<div class='saiv-thinking-item'><div class='saiv-thinking-title'>{safe_title}</div>"
-                f"<div class='saiv-thinking-text'>{safe_text}</div></div>"
-            )
-
-        if not blocks:
-            return base_text
-
-        body = "".join(blocks)
-        details = (
-            "<details class='saiv-thinking'><summary>🧠 Thinking</summary>"
-            f"<div class='saiv-thinking-body'>{body}</div></details>"
-        )
-        return base_text + "\n" + details
-
-    # ------------------------------------------------------------------
-    # Pulse related utilities
-    # ------------------------------------------------------------------
-
-    def _save_conscious_log(self) -> None:
-        self.conscious_log_path.parent.mkdir(parents=True, exist_ok=True)
-        data_to_save = {
-            "log": self.conscious_log,
-            "pulse_cursors": self.pulse_cursors,
-            "pulse_cursor_format": "seq",
-            "pulse_indices": self.pulse_cursors,
-        }
-        self.conscious_log_path.write_text(json.dumps(data_to_save, ensure_ascii=False), encoding="utf-8")
-
-    def _mark_entry(self, building_id: str) -> None:
-        """Mark the latest building message sequence at the moment of entry.
-        First pulse after entry will read only messages after this point.
-        """
-        try:
-            hist = self.history_manager.building_histories.get(building_id, [])
-            last_seq = 0
-            if hist:
-                try:
-                    last_seq = int(hist[-1].get("seq", len(hist)))
-                except (TypeError, ValueError):
-                    last_seq = len(hist)
-            self.entry_markers[building_id] = last_seq
-            prior_cursor = self.pulse_cursors.get(building_id, 0)
-            self.pulse_cursors[building_id] = max(prior_cursor, last_seq)
-            logging.debug("[entry] entry marker set: %s -> %d (prev_cursor=%d)", building_id, last_seq, prior_cursor)
-        except Exception:
-            pass
-
-    def register_entry(self, building_id: str) -> None:
-        """
-        Public hook used when an external manager moves this persona into a building.
-        Resets the perception window so we only ingest messages that happen after arrival.
-        """
-        self._mark_entry(building_id)
 
     def run_pulse(self, occupants: List[str], user_online: bool = True, decision_model: Optional[str] = None) -> List[str]:
         """Execute one autonomous pulse cycle."""
